requirements.txt
phidata 
python-dotenv
yfinance
packaging
duckduckgo-search
fastapi
uvicorn
streamlit
google-generativeai
youtube-transcript-api

